GApps: only clean if making Banks
Open GApps includes their own cleaning functions and this screws with pulling down the new apks so we will omit it for them.

+++ gapps.sh
@@ -125,15 +125,16 @@ START=$( TZ=MST date +%s )
 cd ${SOURCE_DIR}
 
 
+if [[ "${TYPE}" == "banks" ]]; then
+   ############
+   # CLEAN UP #
+   ############
 
-############
-# CLEAN UP #
-############
+   echoText "CLEANING UP REPO"
 
-echoText "CLEANING UP REPO"
-
-git reset --hard origin/${BRANCH}
-git clean -f -d -x
+   git reset --hard origin/${BRANCH}
+   git clean -f -d -x
+fi
 
 
 #####################
